Extract lovelace to it's own component
