+++ homeassistant/components/frontend/__init__.py
@@ -30,7 +30,7 @@ REQUIREMENTS = ['home-assistant-frontend==20180924.0']
 
 DOMAIN = 'frontend'
 DEPENDENCIES = ['api', 'websocket_api', 'http', 'system_log',
-                'auth', 'onboarding']
+                'auth', 'onboarding', 'lovelace']
 
 CONF_THEMES = 'themes'
 CONF_EXTRA_HTML_URL = 'extra_html_url'
@@ -107,10 +107,6 @@ WS_TYPE_GET_TRANSLATIONS = 'frontend/get_translations'
 SCHEMA_GET_TRANSLATIONS = websocket_api.BASE_COMMAND_MESSAGE_SCHEMA.extend({
     vol.Required('type'): WS_TYPE_GET_TRANSLATIONS,
     vol.Required('language'): str,
-})
-WS_TYPE_GET_LOVELACE_UI = 'frontend/lovelace_config'
-SCHEMA_GET_LOVELACE_UI = websocket_api.BASE_COMMAND_MESSAGE_SCHEMA.extend({
-    vol.Required('type'): WS_TYPE_GET_LOVELACE_UI,
 })
 
 
@@ -208,9 +204,6 @@ async def async_setup(hass, config):
     hass.components.websocket_api.async_register_command(
         WS_TYPE_GET_TRANSLATIONS, websocket_get_translations,
         SCHEMA_GET_TRANSLATIONS)
-    hass.components.websocket_api.async_register_command(
-        WS_TYPE_GET_LOVELACE_UI, websocket_lovelace_config,
-        SCHEMA_GET_LOVELACE_UI)
     hass.http.register_view(ManifestJSONView)
 
     conf = config.get(DOMAIN, {})
@@ -530,28 +523,3 @@ def websocket_get_translations(hass, connection, msg):
         ))
 
     hass.async_add_job(send_translations())
-
-
-def websocket_lovelace_config(hass, connection, msg):
-    """Send lovelace UI config over websocket config."""
-    async def send_exp_config():
-        """Send lovelace frontend config."""
-        error = None
-        try:
-            config = await hass.async_add_job(
-                load_yaml, hass.config.path('ui-lovelace.yaml'))
-            message = websocket_api.result_message(
-                msg['id'], config
-            )
-        except FileNotFoundError:
-            error = ('file_not_found',
-                     'Could not find ui-lovelace.yaml in your config dir.')
-        except HomeAssistantError as err:
-            error = 'load_error', str(err)
-
-        if error is not None:
-            message = websocket_api.error_message(msg['id'], *error)
-
-        connection.send_message_outside(message)
-
-    hass.async_add_job(send_exp_config())
